Fix some warnings.

+++ stats.cpp
@@ -60,10 +60,10 @@ std::ostream& Health::print(std::ostream& out) const {
 }
 
 
-Strength::Strength(size_t strength_) : strength(strength_) { }
+Strength::Strength(size_t strength_) : strength(static_cast<int>(strength_)) { }
 
 void Strength::increase(size_t value) {
-  strength += value;
+  strength += static_cast<int>(value);
 }
 
 size_t Strength::get() const {
@@ -89,10 +89,10 @@ void Experience::increase(size_t value) {
   }
 }
 
-Speed::Speed(size_t speed_) : speed(speed_) { }
+Speed::Speed(size_t speed_) : speed(static_cast<int>(speed_)) { }
 
 void Speed::increase(size_t value) {
-  speed += value;
+  speed += static_cast<int>(value);
 }
 
 size_t Speed::get() const {
